Move common items to top of integration tests.
